Register commands from Jarred Classes
Functionality to register methods from classes loaded from Jars in in
WorldEdit's data directory. This allows for easy deployment of Java
based commands without rebuilding and reinstalling WorldEdit. The
jarred-commands directory is expected to be
(platformWorkingDirectory)/plugins/WorldEdit. Actual registration is the
reflection+annotation technique used for compiled-in command
registration. Currently, registration is only done at startup, but
dynamically loading commands on a running server is an expected
enhancement.